+++ client/app.json
@@ -6,7 +6,7 @@
     "platforms": ["ios", "android"],
     "version": "1.0.0",
     "orientation": "portrait",
-    "icon": "./assets/images/logo-sm.png",
+    "icon": "./assets/images/logo-sm-black.png",
     "splash": {
       "image": "./assets/images/splash.png",
       "resizeMode": "contain",
