--- client/src/navigator/Drawer/index.js
@@ -1,3 +1,0 @@
-import Drawer from './Drawer'
-
-export default Drawer
