+++ client/src/slices/auth.slice.js
@@ -1,0 +1,34 @@
+/* eslint-disable no-param-reassign */
+import { createSlice } from '@reduxjs/toolkit'
+
+/*
+  Here is an example slice. A "slice" is a collection of Redux reducer logic and
+  actions for a single feature in your app, typically defined together in a single file.
+*/
+
+// ------------------------------------
+// Constants
+// ------------------------------------
+
+const initialState = {
+    loggedIn: false,
+}
+
+// ------------------------------------
+// Slice
+// ------------------------------------
+
+const authSlice = createSlice({
+    name: 'auth',
+    initialState,
+    reducers: {
+        authenticate: (state, { payload }) => {
+            state.loggedIn = payload.loggedIn
+        },
+    },
+})
+
+export const { action } = authSlice
+export const { authenticate } = authSlice.actions
+
+export default authSlice.reducer
